toncenter WIP

+++ toncenter/src/main/java/org/ton/ton4j/toncenter/model/ConsensusBlockResponse.java
@@ -13,5 +13,5 @@ public class ConsensusBlockResponse {
     private Long consensusBlock;
     
     @SerializedName("timestamp")
-    private Long timestamp;
+    private Double timestamp;
 }
